refactor(git): move library to follow conventions
