+++ libs/schema-registry/src/lib/create-pr-failure/value.ts
@@ -1,4 +1,4 @@
-import { EnumGitProvider } from "@amplication/git-utils";
+import { EnumGitProvider } from "@amplication/util/git";
 import { IsEnum, IsOptional, IsString } from "class-validator";
 
 export class Value {
