+++ libs/util/git/jest.config.ts
@@ -1,13 +1,13 @@
 /* eslint-disable */
 export default {
-  displayName: "amplication-git-utils",
-  preset: "../../jest.preset.js",
+  displayName: "@amplication/util/git",
+  preset: "../../../jest.preset.js",
   testEnvironment: "node",
   transform: {
     "^.+\\.[tj]sx?$": "babel-jest",
   },
   moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
-  coverageDirectory: "../../coverage/packages/amplication-git-utils",
+  coverageDirectory: "../../../coverage/libs/util/git",
   coverageThreshold: {
     global: {
       branches: 31.25,
