+++ libs/util/git/src/errors/NoCommitOnBranch.ts
