+++ libs/util/git/src/index.ts
@@ -1,0 +1,3 @@
+export { GitClientService } from "./git.service";
+export {} from "./git.constants";
+export * from "./types";
