+++ libs/util/git/src/providers/bitbucket/errors.ts
