+++ libs/util/git/src/providers/bitbucket/requests.ts
@@ -1,5 +1,4 @@
 import fetch from "node-fetch";
-import { CustomError } from "../../utils/custom-error";
 import {
   Account,
   OAuth2,
