+++ libs/util/git/src/utils/amplication-ignore-manger.ts
