+++ libs/util/git/src/utils/is-folder-empty/index.ts
