+++ libs/util/git/src/utils/is-folder-empty/is-folder-empty.spec.ts
