+++ libs/util/git/tests/providers/git-service-factory.spec.ts
@@ -1,6 +1,6 @@
 import { EnumGitProvider } from "../../src/types";
 import { GithubService } from "../../src/providers/github/github.service";
-import { GitFactory } from "../../src/providers/git-factory";
+import { GitFactory } from "../../src/git-factory";
 import { mock } from "jest-mock-extended";
 
 describe.skip("GitFactory", () => {
