+++ packages/amplication-git-pull-request-service/src/diff/diff.service.ts
@@ -5,7 +5,7 @@ import { compare } from "dir-compare";
 import { sync } from "fast-glob";
 import { existsSync, readFileSync } from "fs";
 import { normalize } from "path";
-import { File } from "@amplication/git-utils";
+import { File } from "@amplication/util/git";
 import { mapDiffSetToPrModule } from "./diffset-mapper";
 import { BuildPathFactory } from "./build-path-factory";
 import { deleteFilesVisitor } from "./delete-files";
