--- packages/amplication-git-utils/src/index.ts
@@ -1,2 +1,0 @@
-export { GitClientService } from "./providers/git.service";
-export * from "./types";
