+++ tsconfig.base.json
@@ -23,8 +23,6 @@
         "packages/amplication-code-gen-types/src/*"
       ],
       "@amplication/code-gen-utils": ["libs/util/code-gen-utils/src/index.ts"],
-      "@amplication/git-utils": ["packages/amplication-git-utils/src/index.ts"],
-      "@amplication/git-utils/*": ["packages/amplication-git-utils/src/*"],
       "@amplication/kafka/*": ["packages/amplication-kafka/src/*"],
       "@amplication/nest-logger-module/*": [
         "packages/amplication-nest-logger-module/src/*"
@@ -38,6 +36,7 @@
       "@amplication/ui/design-system/icons": [
         "libs/ui/design-system/src/lib/icons"
       ],
+      "@amplication/util/git": ["libs/util/git/src/index.ts"],
       "@amplication/util/json": ["libs/util/json/src/index.ts"],
       "@amplication/util/kafka": ["libs/util/kafka/src/index.ts"],
       "@amplication/util/logging": ["libs/util/logging/src/index.ts"],
