+++ workspace.json
@@ -8,7 +8,6 @@
     "amplication-git-pull-request-service": "packages/amplication-git-pull-request-service",
     "amplication-git-pull-service": "ee/packages/amplication-git-pull-service",
     "amplication-git-push-webhook-service": "ee/packages/amplication-git-push-webhook-service",
-    "amplication-git-utils": "packages/amplication-git-utils",
     "amplication-plugin-api": "packages/amplication-plugin-api",
     "amplication-prisma-db": "packages/amplication-prisma-db",
     "amplication-server": "packages/amplication-server",
@@ -19,6 +18,7 @@
     "ui-design-system": "libs/ui/design-system",
     "ui-design-system-e2e": "packages/ui-design-system-e2e",
     "util-code-gen-utils": "libs/util/code-gen-utils",
+    "util-git": "libs/util/git",
     "util-json": "libs/util/json",
     "util-kafka": "libs/util/kafka",
     "util-logging": "libs/util/logging",
